Cleanup of autorestart, daemonization and lock file checking logic. (#43)
* Cleanup of autorestart, daemonization and lock file checking logic.

Basically, we have to setup autorestart before any network connection is opened (or it will not operate properly after restart), i.e. as early as possible, right after command-line option parsing. But the daemonization have to be even earlier, and lock file checking have to be even before daemonization. So all of this is moved upward in the operation queue. And some device daemons that manually implemented early daemonization (as their low-level kernel drivers dislike forking etc) are cleaned up as they do not need to bother with it anymore.

* Make last commit more pre-CXX11 friendly.