Extend D50's mount driver functionality
Adding path-tracing and selecting functionality, implementing our
original and novel method of performing a gradual movements, named
"Successive safe points" strategy. Everything is quite generic and
prepared for complex movements, however the findPath () function
utilizes only very basic principles for finding the way now - but it
can be easily extended in a the future.

A smarter way of crossing the backlash was also implemented, as well as
other minor changes/repairs.