rts2-newtarget: Create targets with no coordinates
Sometimes we want to deal with scripts that will move the telescope to a position we don't know beforehand, hence, it's useful to have RTS2 don't move the telescope when executing the target.

This commit allows us to create targets with null coordinates. This behaviour is not problematic for RTS2 since the ecpg code has indicator variables that will set the position to NAN (do nothing) if tar_ra or tar_dec is null.

Signed-off-by: Eduardo Alonso <eduardo.alonso@quasarsr.com>