sql: targets: Don't use 0 for targets with undefined position
Some target types, such as DarkTarget, will prevent the telescope from moving no matter which coordinates the target specifies. For some other targets, such as target 3 (Focusing Frames) we don't want the telescope to move to 00:00:00 +00:00:00, although it does until the focusing script kicks in:

EXEC executing now target Focusing frames(#3) at RA DEC 00:00:00.000 +00:00:00.00
T0 moving from (...) to 00:00:00.000 +000:00:00.00 (altaz +026:24:16.86 +176:14:00.00)
T0 moving from (..) to 23:55:29.804 +063:31:42.24 (altaz +089:00:00.00 +090:00:00.00)

Moreover, I find this behaviour a bit inconsistent, since DarkTarget sets the position to NAN but ConstTarget and CalibrationTarget set it to 0 (which is a valid position in the sky).

In the focusing example I highlighted earlier, it's a OportunityTarget, so it simply moves until the script instructs to move to another position.

Instead of having target 3 as a CalibrationTarget, I propose setting the position of these objects to null, instead of zero.

Signed-off-by: Eduardo Alonso <eduardo.alonso@quasarsr.com>