Don't try to close the dome if already closed after bad weather
According to isClosed docs: "It is called also outside of the closing sequence, to check if dome is closed when bad weather arrives."

Right now, I get a lot of "closeDomeWeather: domeCloseStart" spam in the logs when it's daytime and the weather station reports bad weather.

This commit avoids all that spam and will only log if it's not closed.